[PrimTorch] Reference for matrix_norm
As per title. I corrected a thing or two from my previous implementation
to make for better errors in some weird edge-cases and have a more clear
understanding of when does this function support low_precision types and
when it doesn't.

We also use the optimisation for bfloat16 within `vector_norm` within
this function.

ghstack-source-id: cf281dd5428040c59d04432424850303fe40f34f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81113